refactor: Made IInnerRender its own file, used Box3
